set JavaDoc, remove method, rename enum
