+++ src/main/java/fr/techad/edc/popover/AppConfig.java
@@ -1,0 +1,10 @@
+package fr.techad.edc.popover;
+
+public class AppConfig {
+
+    private static String viewerDesktopPath = "";
+
+    public static String getViewerDesktopPath() {
+        return viewerDesktopPath;
+    }
+}
