--- src/main/java/fr/techad/edc/popover/internal/model/HelpViewer.java
@@ -1,5 +1,0 @@
-package fr.techad.edc.popover.internal.model;
-
-public enum HelpViewer {
-    BROWSER, DESKTOP_VIEWER, INTERNAL_BROWSER;
-}
