+++ src/main/java/fr/techad/edc/popover/model/HelpViewer.java
@@ -1,0 +1,5 @@
+package fr.techad.edc.popover.model;
+
+public enum HelpViewer {
+    SYSTEM_BROWSER, EDC_DESKTOP_VIEWER, EMBEDDED_VIEWER;
+}
